Core: SDHC: Genericize the host controller (ugly).
An ugly refactor, but this has divorced the host controller from a
specific device (SD Card). Hopefully this means the host controller will
be able to support both SD memory cards and the WiFi chip (not yet
implemented).

NOTE: This doesn't mean 2 devices on the same host controller, it's
still limited to 1 device per even though the spec allows for up to 8.
This is fine as the Wii has 2 host controllers, with one device per.