we have gdb at home
seriously this protocol is terrible, there are no provisions for the
type of endianness conversion we have to do. This is basically a
non-starter. The new plan is to trash GDB altogether and build our own
thing. This commit reworks the gdb support so far into a more generic
REST-style HTTP service. I will build a new interface in Python, this
will also be a test of removing the brittle Unix Domain Socket based
client in favor of HTTP which is much more flexible.